Tests: moved Stream tests, rewrote to use MemoryStream instead of File
