+++ Solutions/Tests.App/Common.Lib.Test.vcxproj
@@ -24,20 +24,31 @@
     <ClCompile Include="..\..\Common\test\gfxdef_test.cpp" />
     <ClCompile Include="..\..\Common\test\math_test.cpp" />
     <ClCompile Include="..\..\Common\test\memory_test.cpp" />
+    <ClCompile Include="..\..\Common\test\stream_test.cpp" />
     <ClCompile Include="..\..\Common\test\string_test.cpp" />
     <ClCompile Include="..\..\Common\test\version_test.cpp" />
+    <ClCompile Include="..\..\Common\util\alignedstream.cpp" />
+    <ClCompile Include="..\..\Common\util\datastream.cpp" />
+    <ClCompile Include="..\..\Common\util\memorystream.cpp" />
+    <ClCompile Include="..\..\Common\util\proxystream.cpp" />
     <ClCompile Include="..\..\Common\util\stream.cpp" />
     <ClCompile Include="..\..\Common\util\string.cpp" />
     <ClCompile Include="..\..\Common\util\string_compat.c" />
+    <ClCompile Include="..\..\Common\util\string_utils.cpp" />
     <ClCompile Include="..\..\Common\util\version.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Common\gfx\gfx_def.h" />
+    <ClInclude Include="..\..\Common\util\alignedstream.h" />
+    <ClInclude Include="..\..\Common\util\datastream.h" />
     <ClInclude Include="..\..\Common\util\math.h" />
     <ClInclude Include="..\..\Common\util\memory.h" />
+    <ClInclude Include="..\..\Common\util\memorystream.h" />
+    <ClInclude Include="..\..\Common\util\proxystream.h" />
     <ClInclude Include="..\..\Common\util\stream.h" />
     <ClInclude Include="..\..\Common\util\string.h" />
     <ClInclude Include="..\..\Common\util\string_compat.h" />
+    <ClInclude Include="..\..\Common\util\string_utils.h" />
     <ClInclude Include="..\..\Common\util\version.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
